[ISSUE #4013] rename the connect sdk module (#4014)
* [ISSUE #4013] rename the connect sdk module

* rename to open connect
