+++ eventmesh-connectors/sink-connector-rocketmq/build.gradle
@@ -33,7 +33,7 @@ List rocketmq = [
 ]
 
 dependencies {
-    implementation project(":eventmesh-connector-sdks:eventmesh-connector-sdk-java")
+    implementation project(":eventmesh-openconnect:eventmesh-openconnect-java")
     implementation rocketmq
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
