+++ settings.gradle
@@ -23,7 +23,7 @@ include 'eventmesh-common'
 include 'eventmesh-starter'
 include 'eventmesh-examples'
 include 'eventmesh-spi'
-include 'eventmesh-connector-sdks:eventmesh-connector-sdk-java'
+include 'eventmesh-openconnect:eventmesh-openconnect-java'
 include 'eventmesh-connectors:source-connector-rocketmq'
 include 'eventmesh-connectors:sink-connector-rocketmq'
 include 'eventmesh-storage-plugin:eventmesh-storage-api'
